fix(pricing): address all 7 review findings from PR #166
Blockers:
- apply_recommendation: guard payload["product_id"] access; grouped
  category recs now just mark status=applied (no bulk price write)
- SensitivityCalcResponse: Decimal → float so margin_pct serialises as
  a JSON number; previously stored as string → NaN in Saved Scenarios

Correctness:
- _generate_price_recommendations: guard denominator ≤ 0 before dividing
  (target_margin ≥ 100 → skip product instead of ZeroDivisionError)
- calculateSensitivity(): change if(sensUnitCostUsd) to > 0 so the field
  is sent when the backend accepts it (falsy-zero was silently dropped)
- ngOnInit inner getAll() now has an error handler so recsGenerating
  is reset to false even when the fetch fails

UX:
- subsidyInsight medium threshold: ratio < 0.5 → ratio <= 0.5 so a
  50/50 above/below split is not labelled "Well-Balanced"
- buildCrossSubsidy: cap Cartesian pairs at 500 to bound filter cost

+++ backend/src/pricing/schemas.py
@@ -194,15 +194,15 @@ class SensitivityCalcRequest(BaseModel):
 
 
 class SensitivityCalcResponse(BaseModel):
-    unit_cost_usd: Decimal
-    fx_rate: Decimal
-    landed_cost_ngn: Decimal
-    selling_price: Decimal
-    margin_pct: Decimal
+    unit_cost_usd: float
+    fx_rate: float
+    landed_cost_ngn: float
+    selling_price: float
+    margin_pct: float
     quantity: int
-    total_revenue: Decimal
-    total_cost: Decimal
-    gross_profit: Decimal
+    total_revenue: float
+    total_cost: float
+    gross_profit: float
 
 
 # ---------------------------------------------------------------------------
